Clean up integration tests
