+++ tests_integration/__tests__/early-exit.js
@@ -2,9 +2,6 @@
 
 const prettier = require("../../tests_config/require_prettier");
 const runPrettier = require("../runPrettier");
-const constant = require("../../src/cli/constant");
-const util = require("../../src/cli/util");
-const arrayify = require("../../src/utils/arrayify");
 
 describe("show version with --version", () => {
   runPrettier("cli/with-shebang", ["--version"]).test({
@@ -44,34 +41,6 @@ describe(`show detailed usage with plugin options (manual resolution)`, () => {
     "--parser=bar"
   ]).test({
     status: 0
-  });
-});
-
-arrayify(
-  Object.assign(
-    {},
-    util.createDetailedOptionMap(
-      prettier.getSupportInfo(null, {
-        showDeprecated: true,
-        showUnreleased: true,
-        showInternal: true
-      }).options
-    ),
-    util.normalizeDetailedOptionMap(constant.options)
-  ),
-  "name"
-).forEach(option => {
-  const optionNames = [
-    option.description ? option.name : null,
-    option.oppositeDescription ? `no-${option.name}` : null
-  ].filter(Boolean);
-
-  optionNames.forEach(optionName => {
-    describe(`show detailed usage with --help ${optionName}`, () => {
-      runPrettier("cli", ["--help", optionName]).test({
-        status: 0
-      });
-    });
   });
 });
