+++ tests_integration/__tests__/help-options.js
@@ -1,0 +1,42 @@
+"use strict";
+
+const semver = require("semver");
+const isOldNode = semver.parse(process.version).major <= 4;
+if (isOldNode) {
+  test.skip();
+  return;
+}
+
+const prettier = require("../../tests_config/require_prettier");
+const runPrettier = require("../runPrettier");
+const constant = require("../../src/cli/constant");
+const util = require("../../src/cli/util");
+const arrayify = require("../../src/utils/arrayify");
+
+arrayify(
+  Object.assign(
+    {},
+    util.createDetailedOptionMap(
+      prettier.getSupportInfo(null, {
+        showDeprecated: true,
+        showUnreleased: true,
+        showInternal: true
+      }).options
+    ),
+    util.normalizeDetailedOptionMap(constant.options)
+  ),
+  "name"
+).forEach(option => {
+  const optionNames = [
+    option.description ? option.name : null,
+    option.oppositeDescription ? `no-${option.name}` : null
+  ].filter(Boolean);
+
+  optionNames.forEach(optionName => {
+    describe(`show detailed usage with --help ${optionName}`, () => {
+      runPrettier("cli", ["--help", optionName]).test({
+        status: 0
+      });
+    });
+  });
+});
