+++ tests_integration/__tests__/util-shared.js
@@ -1,5 +1,12 @@
 "use strict";
 
+const semver = require("semver");
+const isOldNode = semver.parse(process.version).major <= 4;
+if (isOldNode) {
+  test.skip();
+  return;
+}
+
 const sharedUtil = require("../../src/common/util-shared");
 
 test("shared util has correct structure", () => {
