+++ tests_integration/plugins/extensions/plugin.js
@@ -1,6 +1,7 @@
 "use strict";
 
-const concat = require("../../../src/doc").builders.concat;
+const prettier = require("../../../tests_config/require_prettier");
+const concat = prettier.doc.builders.concat;
 
 module.exports = {
   languages: [
